requirements.txt
aiogram
psycopg[binary]
python-dotenv
qrcode
Pillow